Extracted script module
