+++ src/app/component/admin/admin.module.ts
@@ -4,8 +4,6 @@ import { GetRouteConfigForPage } from "@helpers/page/pageRouting";
 import { SharedModule } from "@shared/shared.module";
 import { adminRoute } from "./admin.menus";
 import { AdminDashboardComponent } from "./dashboard/dashboard.component";
-import { AdminScriptsComponent } from "./scripts/list/list.component";
-import { AdminScriptsNewComponent } from "./scripts/new/new.component";
 import { AdminTagGroupsDeleteComponent } from "./tag-group/delete/delete.component";
 import { AdminTagGroupsEditComponent } from "./tag-group/edit/edit.component";
 import { AdminTagGroupsComponent } from "./tag-group/list/list.component";
@@ -15,11 +13,11 @@ import { AdminTagsEditComponent } from "./tags/edit/edit.component";
 import { AdminTagsComponent } from "./tags/list/list.component";
 import { AdminTagsNewComponent } from "./tags/new/new.component";
 import { AdminUserListComponent } from "./users/list/list.component";
+import { ScriptsModule } from "./scripts/scripts.module";
 
+const modules = [ScriptsModule];
 const components = [
   AdminDashboardComponent,
-  AdminScriptsComponent,
-  AdminScriptsNewComponent,
   AdminTagGroupsComponent,
   AdminTagGroupsDeleteComponent,
   AdminTagGroupsEditComponent,
@@ -34,7 +32,7 @@ const routes = adminRoute.compileRoutes(GetRouteConfigForPage);
 
 @NgModule({
   declarations: components,
-  imports: [SharedModule, RouterModule.forChild(routes)],
-  exports: [RouterModule, ...components],
+  imports: [SharedModule, RouterModule.forChild(routes), ...modules],
+  exports: [RouterModule, ...components, ...modules],
 })
 export class AdminModule {}
