+++ src/app/component/admin/scripts/scripts.menus.ts
@@ -1,0 +1,31 @@
+import { defaultNewIcon, isAdminPredicate } from "src/app/app.menus";
+import { Category, MenuRoute } from "src/app/interfaces/menusInterfaces";
+import { adminDashboardMenuItem } from "../admin.menus";
+
+export const adminScriptsRoute = adminDashboardMenuItem.route.addFeatureModule(
+  "scripts"
+);
+
+export const adminScriptsCategory: Category = {
+  icon: ["fas", "scroll"],
+  label: "Scripts",
+  route: adminScriptsRoute,
+};
+
+export const adminScriptsMenuItem = MenuRoute({
+  icon: ["fas", "scroll"],
+  label: "Scripts",
+  route: adminScriptsCategory.route,
+  tooltip: () => "Manage custom scripts",
+  parent: adminDashboardMenuItem,
+  predicate: isAdminPredicate,
+});
+
+export const adminNewScriptsMenuItem = MenuRoute({
+  icon: defaultNewIcon,
+  label: "New Script",
+  route: adminScriptsMenuItem.route.add("new"),
+  tooltip: () => "Create a new script",
+  parent: adminScriptsMenuItem,
+  predicate: isAdminPredicate,
+});
