+++ src/app/component/admin/scripts/scripts.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from "@angular/core";
+import { RouterModule } from "@angular/router";
+import { GetRouteConfigForPage } from "src/app/helpers/page/pageRouting";
+import { SharedModule } from "../../shared/shared.module";
+import { AdminScriptsComponent } from "./list/list.component";
+import { AdminScriptsNewComponent } from "./new/new.component";
+import { adminScriptsRoute } from "./scripts.menus";
+
+const components = [AdminScriptsComponent, AdminScriptsNewComponent];
+const routes = adminScriptsRoute.compileRoutes(GetRouteConfigForPage);
+
+@NgModule({
+  declarations: components,
+  imports: [SharedModule, RouterModule.forChild(routes)],
+  exports: [RouterModule, ...components],
+})
+export class ScriptsModule {}
